% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_of_all_classes.R
\name{frequency_of_all_classes}
\alias{frequency_of_all_classes}
\title{Create bar plots of frequency of tail types for tested strains}
\usage{
frequency_of_all_classes(
  listofnames = strainslist,
  data = dataList2,
  keepMIX = F,
  plotcolours = c("#6C91E8", "#64F069", "#C04EDC", "#D5B54B", "#F23C28"),
  same_scale = T,
  plotwidth,
  plotheight,
  returnplots = T
)
}
\arguments{
\item{listofnames}{The output of gw3RACE::pair_replicates(). By default = strainslist.}

\item{data}{The output of gw3RACE::tail_summary(). By default = dataList2.}

\item{keepMIX}{Whether mixed_tailed class should be plotted or not. By default = F.}

\item{plotcolours}{A vector of 5 colours. By default = c("#6C91E8", "#64F069", "#C04EDC", "#D5B54B", "#F23C28").}

\item{same_scale}{Whether all generated plots should have the same limits of Y axis. By default = T.}

\item{plotwidth}{Numeric value of width of plots that is being passed to pdf function.}

\item{plotheight}{Numeric value of height of plots that is being passed to pdf function.}

\item{returnplots}{If TRUE function will create bar plots (separate for each strain) within new folder ../types/, if you don't like the plots and want to create your own, type FALSE and function will return a list of data frames with data ready to be plotted. By default = T.}
}
\value{
Either a list of data frames or .pdf files with barplots
}
\description{
frequency_of_all_classes() creates bar plots of mean frequencies of all classes of RNA tails for every strain within newly created directory ../types.
}
\examples{
frequency_of_all_classes(plotwidth = 5, plotheight = 3)
}
