% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_of_class.R
\name{frequency_of_class}
\alias{frequency_of_class}
\title{Create bar plot of frequency of given tail type among tested strains}
\usage{
frequency_of_class(
  listofnames = strainslist,
  data = dataList2,
  tailclass,
  nameofcontrol,
  returnplot = T
)
}
\arguments{
\item{listofnames}{The output of gw3RACE::pair_replicates(). By default = strainslist.}

\item{data}{The output of gw3RACE::tail_summary(). By default = dataList2.}

\item{tailclass}{Class of a RNA tails that is to be plotted (choose one these: "mixed_tail" "no_tail" "oligoU" "polyA" "polyAU")}

\item{nameofcontrol}{Name of strain that served as a control in the experiment, must be exactly the same as in the output of gw3RACE::pair_replicates().}

\item{returnplot}{If TRUE function will return a bar plot, if you don't like the plot and want to create your own, type FALSE and function will return a data frame with data ready to be plotted}
}
\description{
frequency_of_class() creates a bar plot of mean frequency of a class of RNA tails for every strain
}
\examples{
frequency_of_class(nameofcontrol = "Wild_type", tailclass = "polyAU")
}
