% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepcommongenes.R
\name{keepcommongenes}
\alias{keepcommongenes}
\title{Keep the genes that surpassed the cutoff for every single sample in your experiment}
\usage{
keepcommongenes(data, querry = dataList2)
}
\arguments{
\item{data}{A data frame stored on the list that is an output of gw3RACE::data_fix() or gw3RACE::tail_summary(), generally it is recommended to use this function with lapply()}

\item{querry}{Output of gw3RACE::tail_summary(). By default = dataList2.}
}
\value{
List of data frames
}
\description{
keepcommongenes() filters data for genes that are present in all samples
}
\examples{
dataList2 <- lapply (dataList2, keepcommongenes)
dataList <- lapply (dataList, keepcommongenes)
}
