% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nofReads.R
\name{nofReads}
\alias{nofReads}
\title{Plot of number of reads/genes in sample}
\usage{
nofReads(listofcsv)
}
\arguments{
\item{listofcsv}{A list of data frames, output of gw3RACE::data_fix() or gw3RACE::tail_summary()}
}
\value{
Barplot
}
\description{
Plot number of reads in data frames stored on list with gw3RACE::nofReads()
}
\examples{
nofReads(dataList)
}
