% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_replicates.R
\name{pair_replicates}
\alias{pair_replicates}
\title{Teach R to recognize replicates of independent samples in Genome-wide 3' RACE}
\usage{
pair_replicates(data, grep1 = "_clone\\\\d")
}
\arguments{
\item{data}{A list that is an output of gw3RACE::data_fix()}

\item{grep1}{an argument passed to gsub(), regular expression that can be used to erase the indication of replication of experiment from name of data table stored on the list. By default it is "_clone\\\d"}
}
\value{
List of vectors
}
\description{
pair_replicates() creates list of two character vectors, first contains names of used strains, second contains expressions that can be used to grep the said strains from the list of data frames
}
\examples{
strainslist <- pair_replicates(dataList)
}
