% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_of_tail_length.R
\name{plot_of_tail_length}
\alias{plot_of_tail_length}
\title{Create violin plot of RNA tails length}
\usage{
plot_of_tail_length(
  listofnames = strainslist,
  data = dataList2,
  nameofcontrol,
  returnplot = T
)
}
\arguments{
\item{listofnames}{The output of gw3RACE::pair_replicates(). By default = strainslist.}

\item{data}{The output of gw3RACE::tail_summary(). By default = dataList2.}

\item{nameofcontrol}{Name of strain that served as a control in the experiment, must be exactly the same as in the output of gw3RACE::pair_replicates()}

\item{returnplot}{if TRUE function will return a violin plot, if you don't like the plot and want to create your own, type FALSE and function will return a data frame with data ready to be plotted. By defult = T.}
}
\value{
Either a data frame or a violin plot based on it
}
\description{
plot_of_tail_length() creates violin plot of mean tail lengths per gene measured in Genome-wide 3' RACE for all strains
}
\examples{
plot_of_tail_length(nameofcontrol = "Wild_type")
}
