% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readallcsvinfolder.R
\name{readallcsvinfolder}
\alias{readallcsvinfolder}
\title{Upload all .csv files in one folder to R list}
\usage{
readallcsvinfolder(
  folder = getwd(),
  columnstokeep = c("gene", "coord_R2", "RNA_type", "tail_from", "tail_len", "tail_type",
    "stop_R2", "distance_to_TES", "rel_distance_to_TES"),
  RNAtype = "mRNA"
)
}
\arguments{
\item{folder}{A path to the folder where you store your .csv files, by default it is your current working directory}

\item{columnstokeep}{A vector of column names that is passed to subset function, by default it's = c("gene", "coord_R2", "RNA_type", "tail_from", "tail_len", "tail_type", "stop_R2", "distance_to_TES", "rel_distance_to_TES")}

\item{RNAtype}{A character or a vector of characters describing RNA types, that you want to keep in your data, must be identical to names used in RNA_type column in original file, by default it's "mRNA"}
}
\value{
List of data frames
}
\description{
readallcsvinfolder() will read all .csv files in given folder and create list of them
}
\examples{
dataList <- readallcsvinfolder()
}
