% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_profiles.R
\name{tail_profiles}
\alias{tail_profiles}
\title{Create histograms of distribution of length of tails of different classes.}
\usage{
tail_profiles(
  listofnames = strainslist,
  data = dataList,
  plotcolours = c("#6C91E8", "#64F069", "#C04EDC"),
  same_scale = T,
  plotwidth,
  plotheight,
  returnplots = T
)
}
\arguments{
\item{listofnames}{The output of gw3RACE::pair_replicates(). By default = strainslist.}

\item{data}{The output of gw3RACE::data_fix(). By default = dataList.}

\item{plotcolours}{A vector of 3 colours. By default = c("#6C91E8", "#64F069", "#C04EDC")}

\item{same_scale}{Whether all generated plots should have the same limits of Y axis. By default = T.}

\item{plotwidth}{Numeric value of width of plots that is being passed to pdf function.}

\item{plotheight}{Numeric value of height of plots that is being passed to pdf function.}

\item{returnplots}{If TRUE function will create histograms (separate for each strain) within new folder ../profiles/, if you don't like the plots and want to create your own, type FALSE and function will return a list of data frames with data ready to be plotted. By default = T.}
}
\value{
Either a list of data frames or .pdf files with histograms
}
\description{
tail_profiles() creates histograms of the frequency of occurrence of individual classes of tails, taking into account their length. New plots are generated within newly created directory ../profiles.
}
\examples{
tail_profiles(plotwidth = 7, plotheight = 5)
}
