% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_summary.R
\name{tail_summary}
\alias{tail_summary}
\title{Get summary of per gene analysis}
\usage{
tail_summary(data, threshold = 20, gen_len)
}
\arguments{
\item{data}{A data frame stored on the list that is an output of gw3RACE::data_fix(), generally it is recommended to use this function with lapply()}

\item{threshold}{A numeric value, minimal number of reads per gene in a sample that qualifies that gene for future analysis, by default it's equal to 20}

\item{gen_len}{A data frame that contains lengths of all transcripts of studied organism (TRANSCRIPTS NOT ORFS), it is used to normalize notail reads to gene length. By default = gen_len.}
}
\value{
List of data frames
}
\description{
tail_summary() summarises data that is stored in the initial data frames so now it holds info on (per gene) mean tail length and frequency of each tail kind
}
\examples{
dataList2 <- lapply(dataList, tail_summary, gen_len = gen_len)
}
